backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
passlib[bcrypt]
bcrypt==4.2.1
python-jose[cryptography]
python-dotenv
pydantic[email]
python-multipart
